django_movie/authentication/urls.py
from django.urls import path, include, re_path
from rest_framework_simplejwt.views import (
    TokenObtainPairView,
    TokenRefreshView,
    TokenVerifyView,
)


urlpatterns = [
    path("djoser/auth/", include('djoser.urls')),
    path("djoser/jwt/auth/", include('djoser.urls.jwt')),
    re_path(r"^auth/", include("djoser.urls.authtoken")),
    path('token/', TokenObtainPairView.as_view(), name='token_obtain_pair'),
    path('token/refresh/', TokenRefreshView.as_view(), name='token_refresh'),
    path('token/verify/', TokenVerifyView.as_view(), name='token_verify'),
]
